codegen/requirements.txt
PyYAML>=5.4
watchdog 